+++ pubspec.yaml
@@ -22,6 +22,8 @@ dependencies:
   flutter_mobx: ^2.0.0
   build_runner: ^2.1.5
   flutter_typeahead: ^3.1.3
+  flutter_localizations:
+    sdk: flutter
 
 dev_dependencies:
   flutter_test:
